check for toast after every query saving to fix flickering spec
