+++ spec/support/pages/page.rb
@@ -26,12 +26,15 @@
 # See COPYRIGHT and LICENSE files for more details.
 #++
 
+require_relative '../toasts/expectations'
+
 module Pages
   class Page
     include Capybara::DSL
     include Capybara::RSpecMatchers
     include RSpec::Matchers
     include OpenProject::StaticRouting::UrlHelpers
+    include Toasts::Expectations
 
     def current_page?
       URI.parse(current_url).path == path
@@ -91,36 +94,6 @@ module Pages
       expect(page).to have_current_path expected_path, wait: 10
     end
 
-    def expect_toast(message:, type: :success)
-      if toast_type == :angular
-        expect(page).to have_selector(".op-toast.-#{type}", text: message, wait: 20)
-      elsif type == :error
-        expect(page).to have_selector(".errorExplanation", text: message)
-      elsif type == :success
-        expect(page).to have_selector(".op-toast.-success", text: message)
-      else
-        raise NotImplementedError
-      end
-    end
-
-    def expect_and_dismiss_toaster(message: nil, type: :success)
-      expect_toast(type:, message:)
-      dismiss_toaster!
-      expect_no_toaster(type:, message:)
-    end
-
-    def dismiss_toaster!
-      page.find('.op-toast--close').click
-    end
-
-    def expect_no_toaster(type: :success, message: nil)
-      if type.nil?
-        expect(page).not_to have_selector(".op-toast")
-      else
-        expect(page).not_to have_selector(".op-toast.-#{type}", text: message)
-      end
-    end
-
     def click_to_sort_by(header_name)
       within '.generic-table thead' do
         click_link header_name
@@ -178,10 +151,6 @@ module Pages
 
     def path
       nil
-    end
-
-    def toast_type
-      :angular
     end
 
     def navigate_to_modules_menu_item(link_title)
